# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ActiveJourneyStatus = typeof ActiveJourneyStatus[keyof typeof ActiveJourneyStatus];


export const ActiveJourneyStatus = {
  CONFIRMED: 'CONFIRMED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
} as const;

export type JourneyLegMode = typeof JourneyLegMode[keyof typeof JourneyLegMode];


export const JourneyLegMode = {
  WALK: 'WALK',
  BUS: 'BUS',
  METRO: 'METRO',
  CAB: 'CAB',
} as const;

export type JourneyLegStatus = typeof JourneyLegStatus[keyof typeof JourneyLegStatus];


export const JourneyLegStatus = {
  PLANNED: 'PLANNED',
  HELD: 'HELD',
  ACTIVE: 'ACTIVE',
  COMPLETED: 'COMPLETED',
  SKIPPED: 'SKIPPED',
} as const;

export interface JourneyLeg {
  id: string;
  mode: JourneyLegMode;
  provider: string;
  from: string;
  to: string;
  durationMinutes: number;
  farePaise: number;
  holdPaise: number;
  status: JourneyLegStatus;
  accent: string;
}

export type ActiveJourneyTicket = {
  id: string;
  code: string;
  expiresInSeconds: number;
};

export interface ActiveJourney {
  id: string;
  from: string;
  to: string;
  status: ActiveJourneyStatus;
  totalFarePaise: number;
  heldPaise: number;
  currentLegIndex: number;
  confirmedAt: string;
  legs: JourneyLeg[];
  ticket: ActiveJourneyTicket;
}

export interface RecentJourney {
  id: string;
  from: string;
  to: string;
  completedAt: string;
  totalFarePaise: number;
  modes: string[];
}

export interface DashboardSummary {
  greeting: string;
  balancePaise: number;
  /** @nullable */
  activeJourney: ActiveJourney | null;
  recentJourneys: RecentJourney[];
  simTime: string;
}

export interface JourneyOption {
  id: string;
  label: string;
  durationMinutes: number;
  arrivalTime: string;
  totalFarePaise: number;
  holdPaise: number;
  confidence: number;
  recommended: boolean;
  legs: JourneyLeg[];
}

export interface JourneyConfirmationInput {
  optionId: string;
}

export interface Wallet {
  balancePaise: number;
  heldPaise: number;
  availablePaise: number;
  currency: string;
}

export type LedgerEntryType = typeof LedgerEntryType[keyof typeof LedgerEntryType];


export const LedgerEntryType = {
  CREDIT: 'CREDIT',
  HOLD: 'HOLD',
  RELEASE: 'RELEASE',
  DEBIT: 'DEBIT',
} as const;

export interface LedgerEntry {
  id: string;
  type: LedgerEntryType;
  label: string;
  amountPaise: number;
  createdAt: string;
  status: string;
}

export interface WalletTopUpInput {
  /** @minimum 100 */
  amountPaise: number;
}

export interface ValidateJourneyLegInput {
  legId: string;
}

export type GetJourneyOptionsParams = {
from: string;
to: string;
departAt?: string;
};