# lib/api-zod/src/generated/types/activeJourney.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActiveJourneyStatus } from './activeJourneyStatus';
import type { ActiveJourneyTicket } from './activeJourneyTicket';
import type { JourneyLeg } from './journeyLeg';

export interface ActiveJourney {
  id: string;
  from: string;
  to: string;
  status: ActiveJourneyStatus;
  totalFarePaise: number;
  heldPaise: number;
  currentLegIndex: number;
  confirmedAt: string;
  legs: JourneyLeg[];
  ticket: ActiveJourneyTicket;
}