# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActiveJourney } from './activeJourney';
import type { RecentJourney } from './recentJourney';

export interface DashboardSummary {
  greeting: string;
  balancePaise: number;
  /** @nullable */
  activeJourney: ActiveJourney | null;
  recentJourneys: RecentJourney[];
  simTime: string;
}