# lib/api-zod/src/generated/types/journeyLeg.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JourneyLegMode } from './journeyLegMode';
import type { JourneyLegStatus } from './journeyLegStatus';

export interface JourneyLeg {
  id: string;
  mode: JourneyLegMode;
  provider: string;
  from: string;
  to: string;
  durationMinutes: number;
  farePaise: number;
  holdPaise: number;
  status: JourneyLegStatus;
  accent: string;
}