# lib/api-zod/src/generated/types/journeyOption.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JourneyLeg } from './journeyLeg';

export interface JourneyOption {
  id: string;
  label: string;
  durationMinutes: number;
  arrivalTime: string;
  totalFarePaise: number;
  holdPaise: number;
  confidence: number;
  recommended: boolean;
  legs: JourneyLeg[];
}