# lib/api-zod/src/generated/types/ledgerEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LedgerEntryType } from './ledgerEntryType';

export interface LedgerEntry {
  id: string;
  type: LedgerEntryType;
  label: string;
  amountPaise: number;
  createdAt: string;
  status: string;
}