# lib/api-zod/src/generated/types/recentJourney.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecentJourney {
  id: string;
  from: string;
  to: string;
  completedAt: string;
  totalFarePaise: number;
  modes: string[];
}